requirements-dev.txt
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
numpy==2.2.6
opencv-python==4.13.0.92
packaging==26.0
pandas==2.3.3
-e file:///D:/intezet/Pico_algae
pillow==12.1.1
pip==26.0.1
pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
rich==14.3.3
six==1.17.0
tqdm==4.67.3
tzdata==2025.3
